Partial Bluetooth Revision
Can now choose which device to pair to from within the app... not pretty
yet, but it works.  Also, if the device is not yet paired, the user is
prompted for the pair-key. No longer need to go to the settings
window...
Still halfway through converting the Bluetooth class into a service...
Working on receiving data next.